Commit 1
The objective of this project is to develop an automated pipeline that extracts football data from public sources, transforms it into a structured format, and loads it into a PostgreSQL database for further use by our data science and data analytics team to enable management make data driven decisions.
This pipeline aims to streamline the process of gathering and processing historical sports data, enabling more accurate predictive models for insight analysis and betting.